backend/requirements.txt
cdifflib
alembic
fastapi
PyMySQL
python-dotenv
SQLAlchemy==1.3.20
SQLAlchemy-Utils
uvicorn
passlib
pyjwt
requests
urllib3
python-geohash
numpy
pystache
python-multipart
python-dateutil
pyyaml
pytz
progress